# src/entitys/project.entity.ts
import { Entity, PrimaryGeneratedColumn, Column, OneToMany } from "typeorm";
// import { TodoList } from "./todoList.entity";
import { Task } from "./task.entity";

@Entity({ name: 'project', schema: 'public' })
export class Project {
    @PrimaryGeneratedColumn("uuid")
    id: string;

    @Column()
    name: string;

    @Column()
    description: string;

    @Column()
    startDate: Date;

    @Column()
    endDate: Date;

    @OneToMany(type => Task, task => task.project)
    task: Task[];
}